+++ spec/rspec/matchers/built_in/negation_spec.rb
@@ -4,11 +4,7 @@ module RSpec::Matchers::BuiltIn
     def configured_matcher_boolean_operators(is_enabled)
       RSpec.configure do |config|
         config.expect_with :rspec do |c|
-          if is_enabled
-            c.enable_matcher_boolean_operators!
-          else
-            c.disable_matcher_boolean_operators!
-          end
+          c.matcher_boolean_operators = is_enabled
         end
       end
     end